Move select_one_where, select_many_where to sqlite.query_helpers
